Landscape function

Create a virtual landscape with different patches.

createLandscape(r, size_pp, n_nf, size_nf)

Arguments

r
A raster object

size_pp
The size of pine plantation expressed as pixel numbers (integer)

n_nf
Number of natural forest patchs to create (integer, between 1 and 5)

size_nf
The size of natural forest patchs (numeric value, between 10 and 500) expressed as number of pixels

Value

A raster object with the landscape configured

Details

a target pine plantation

natural forests

crops

shrublands

The size of the pine plantation and natural forest patchs are specified by size_pp and size_nf. First the pine plantation patch is set with the dimensions specified by size_pp. Then a number of natural forest patchs (n_nf, between 1 and 5) are created with a determined size (size_nf). The crop patches (a random number between 3 and 8) are then established according to the remaining space. Finally, remain pixels are assigned to the shrubland category. The size and proportion of the each patch classes range according to distribution of those ecosystem in Sierra Nevada (SE Spain) (see Pérez-Luque et al. 2014 and Pérez-Luque et al. 2019)

References

Pérez-Luque AJ, Bonet FJ, Pérez-Pérez R, Aspizua R, Lorite J, Zamora R (2014). “Sinfonevada: Dataset of floristic diversity in Sierra Nevada forests (SE Spain).” PhytoKeys, 35, 1--15. doi: 10.3897/phytokeys.35.6363 .

Pérez-Luque AJ, Bonet-García FJ, Zamora Rodríguez R (2019). “Map of ecosystems types in Sierra Nevada mountain (southern Spain).” Pangaea. doi: 10.1594/PANGAEA.91017 .