Seed input by dispersers

Compute the propagule input from each patch to focal pine plantation using three classes of dispersers and different equations.

disper(x, xr, nf_value, pp_value)

Arguments

x
A raster object with the landscape configured

xr
A raster object with richness computed

nf_value
The value of "Natural Forests" class within the raster (default value = 2)

pp_value
The value of "pine plantation" class within the raster (default value = 1)

Details

It computes the propagule input from each patch to focal pine plantation using three classes of dispersers and different equations.

Seed sources: seed diversity in seed source patch, and patch size

Disperser: percentage of each disperser

Landscape configuration

small birds, e.g. European robin (Erithacus rubecula), Sardinian warbler (Sylvia melanocephala)

medium birds, e.g. Eurasian jay (Garrulus glandarius)

mammals, e.g. Red fox (Vulpes vulpes)

Small-sized birds rarely exceed 100 m in distance, and approximately 50% of the seeds are dispersed in the first 50 m.

Medium-sized birds disperse 50% of the seeds over a distance of more than 100 m. The Eurasyan jay shows a dispersin range between 5 and 1000 m for Sierra Nevada mountains (SE Spain). The distance at which the maximum dispersion occurs depends on the target patch, being approximately 400 me when the target patch is a pine plantation.

Mammals disperse in a range from 0 to more than 1500 m, with the dispersion peak at 650 - 700 m. More than 50% of the seeds dispersed by mammals are deposited at distances greater than 495 m.

References

Gómez JM (2003). “Spatial patterns in long-distance dispersal of Quercus ilex acorns by jays in a heterogeneous landscape.” Ecography, 26(5), 573--584. doi: 10.1034/j.1600-0587.2003.03586.x .

González-Varo JP, López-Bao JV, Guitián J (2013). “Functional diversity among seed dispersal kernels generated by carnivorous mammals.” Journal of Animal Ecology, 82(3), 562--571. doi: 10.1111/1365-2656.12024 .

Jordano P, García C, Godoy JA, García-Castaño JL (2007). “Differential contribution of frugivores to complex seed dispersal patterns.” Proceedings of the National Academy of Sciences, 104(9), 3278--3282. doi: 10.1073/pnas.0606793104 , http://www.pnas.org/content/104/9/3278.abstract.

Matías L, Zamora R, Mendoza I, Hódar JA (2010). “Seed dispersal patterns by large frugivorous mammals in a degraded mosaic landscape.” Restoration Ecology, 18(5), 619--627. doi: 10.1111/j.1526-100X.2008.00475.x .

Pons J, Pausas JG (2007). “Acorn dispersal estimated by radio-tracking.” Oecologia, 153(4), 903--911. doi: 10.1007/s00442-007-0788-x .

Zamora R, Hódar JA, Matías L, Mendoza I (2010). “Positive adjacency effects mediated by seed disperser birds in pine plantations.” Ecological Applications, 20(4), 1053--1060. doi: 10.1890/09-0055.1 .