Propagule input into pine plantation by time

It computes the seeds input into focal pine plantation along the time.

disper_time(
  msb,
  mmb,
  mma,
  x,
  xr,
  pp_value,
  per_sb,
  per_mb,
  per_ma,
  propaguleInputBird,
  propaguleInputMammal,
  time_span
)

Arguments

msb
A raster object with small bird contribution on target pine plantation

mmb
A raster object with medium bird contribution on target pine plantation

mma
A raster object with mammal contribution on target pine plantation

x
A raster object with the landscape configured

xr
A raster object with richness computed

pp_value
The value of "Pine plantation" class within the raster (Default value = 1)

per_sb
Percentage of small bird dispersers

per_mb
Percentage of medium bird dispersers

per_ma
Percentage of mammals dispersers

propaguleInputBird
Propagule input by birds (seeds number by year and square meter) (see Zamora et al. 2010)

propaguleInputMammal
Propagule input by mammals (seeds number by year and square meter) (see Zamora et al. 2010)

time_span
Numbers of year simulated (maximum 50)

Details

Compute the propagule input from each natural forest patch to pine plantation target using three types of disperser and different equations. The seed input into focal pine plantation depends on landscape configuration: distance to the seed source and quantity and typology of vegetation patches; and on the composition of the dispersal community. Data about amount of seed introduced into pine plantation in Sierra Nevada mountain (SE Spain) comes from ( Zamora et al. 2010) .

References

Zamora R, Hódar JA, Matías L, Mendoza I (2010). “Positive adjacency effects mediated by seed disperser birds in pine plantations.” Ecological Applications, 20(4), 1053--1060. doi: 10.1890/09-0055.1 .