Set values of Initial Richness

This function computes the initial values of richness values for each patch.

initRichness(r, draster, r_range, treedensity, pastUse, rescale = TRUE)

Arguments

r

A raster object with the configured landscape

draster

A raster object with values of the distance from the target patch to natural forest patchs

r_range

A data frame with three columns: value of land use (integer: 0 = "Other", 1 = "Pine plantation", 2 = "Natural Forests", 3 = "Crop"); lowRich and upRich (lower and upper value of the range of Richness)(see Gómez-Aparicio et al. 2009) .

treedensity

density of the pine plantation (integer)

pastUse

the past land use of the pine plantation (character). One of "Oak", "Shrubland", "Pasture" or "Crop"

rescale

If "TRUE" the results are rescaled (0 = min and 1 = max)

Value

A raster object with values of initial richness for each pixel.

Details

This function computes the initial richness for each land-use categories. The richenss values of each pixel of the focal (target) pine patch depends on:

Stand structure: tree density, patch size, past land-use

Distance to seed source (landscape configuration)

Richenss value for each of the patch classes (i.e. pine plantation, natural forests, shrubland and crops) are calculated considering the range of possible values found on the study area. In this case we use data from Sierra Nevada (southern Spain) See References.

References

Gómez-Aparicio L, Zavala MA, Bonet FJ, Zamora R (2009). “Are pine plantations valid tools for restoring Mediterranean forests? An assessment along abiotic and biotic gradients.” Ecological Applications, 19(8), 2124--2141. doi:10.1890/08-1656.1 .

Mendoza I, Gómez-Aparicio L, Zamora R, Matías L (2009). “Recruitment limitation of forest communities in a degraded Mediterranean landscape.” Journal of Vegetation Science, 20(2), 367--376. doi:10.1111/j.1654-1103.2009.05705.x .

González-Moreno P, Quero JL, Poorter L, Bonet FJ, Zamora R (2011). “Is spatial structure the key to promote plant diversity in Mediterranean forest plantations?” Basic and Applied Ecology, 12(3), 251--259. doi:10.1016/j.baae.2011.02.012 .

Navarro-González I, Pérez-Luque AJ, Bonet FJ, Zamora R (2013). “The weight of the past: land-use legacies and recolonization of pine plantations by oak trees.” Ecological Applications, 23(6), 1267--1276. doi:10.1890/12-0459.1 .

Pérez-Luque AJ, Bonet FJ, Pérez-Pérez R, Aspizua R, Lorite J, Zamora R (2014). “Sinfonevada: Dataset of floristic diversity in Sierra Nevada forests (SE Spain).” PhytoKeys, 35, 1--15. doi:10.3897/phytokeys.35.6363 .

Author

Antonio J Pérez-Luque (ajpelu@gmail.com)