Compute the potential dispersion

Compute the propagule input from each patch to target pine plantation using three classes of disperses and different kernels equations.

Usage

potential_dispersion(
  x,
  nf_value,
  pp_value,
  rich_nf,
  kernel_sbi,
  kernel_mbi,
  kernel_ma,
  seedlim_int,
  seedlim_slope
)

Arguments

x

A raster object with the landscape configured

nf_value

The value of "Natural Forests" class within the raster (default value = 2).

pp_value

The value of "pine plantation" class within the raster (default value = 1).

rich_nf

A raster object with richness of the natural forests

kernel_sbi, kernel_mbi, kernel_ma

The dispersion kernel functions for small birds (kernel_sbi), medium birds (kernel_mbi) and mammals (kernel_ma). See details.

seedlim_int, seedlim_slope

The intercept and slope of the linear relation between the seed limitation in the target plantation and the adjacency of a natural forest patch with the target pine plantation. See details

Value

raster objects with values of potential dispersion for each type of disperser

Details

It computes the propagule input from each patch to focal pine plantation using three classes of disperses and different kernels.

Seed sources: seed diversity in seed source patch, and patch size

Disperses: percentage of each disperser type

Landscape configuration

small birds, e.g. European robin (Erithacus rubecula), Sardinian warbler (Sylvia melanocephala)

medium birds, e.g. Eurasian jay (Garrulus glandarius)

mammals, e.g. Red fox (Vulpes vulpes)

Small-sized birds rarely exceed 100 m in distance, and approximately 50% of the seeds are dispersed in the first 50 m.

Medium-sized birds disperse 50% of the seeds over a distance of more than 100 m. The Eurasyan jay shows a dispersion range between 5 and 1000 m for Sierra Nevada mountains (SE Spain). The distance at which the maximum dispersion occurs depends on the target patch, being approximately 400 m when the target patch is a pine plantation.

Mammals disperse in a range from 0 to more than 1500 m, with the dispersion peak at 650 - 700 m. More than 50% of the seeds dispersed by mammals are deposited at distances greater than 495 m.

According to the disperser type, the function uses by default different dispersion kernels. For small and medium birds a log-normal dispersion kernels were used. Specifically, log-normal density function with mean and standard deviation values of log(51) and log(2) respectively for small-sized birds, and log-normal density function with mean and standard deviation values of log(201) and log(2) for medium-sized birds. This distribution were implemented using the stats::dlnorm() function. For mammals, a combination of log-normal and Weibull dispersion kernel was used. Specifically, from 0 to 400 m of distance from the seed source, a Weibull distribution with a shape and scale parameters of 1.385 and 137 respectively was used (See stats::dweibull()); whereas a log-normal density function with mean and standard deviation values of log(751) and log(1.346) respectively, were applied for distances higher than 401 m from seed source. Different kernels functions could be specified using the parameters kernel_sbi, kernel_mbi, and kernel_ma.

This function also considers the adjacency between each of the natural forest patches and the target pine plantation. The higher the adjacency between the natural forest and the pine plantation, the lower the limitation of the propagule entry dispersed by birds. Zamora et al. (2010) found that the intercept and the slope of the linear relation between the seed limitation (values from 0 to 1) and the adjacency (i.e. percentage of pine-plantation perimeter in contact with native forests) were 0.733 and 0.0039 respectively. Those parameters are included by default, but could be customized using the seedlim_int and seedlim_slope parameters. For each natural forest patch, the adjacency to target pine-plantation is computed. For those patches with adjacency, the potential dispersion by birds increases according a correction factor (see Zamora et al. 2010), that is computed as follows:

\(adj_{fc} = 1 + \frac{\textup{seed Entry} - \textup{seed Entry}_{0}}{\textup{seed Entry}_{100} - \textup{seed Entry}_{0}}\)

where \(\textup{seed Entry} = 1- \textup{seed limitation}\); seed limitation is computed using the seedlim_int and seedlim_slope parameters; \(\textup{seed Entry}_{0}\) and \(\textup{seed Entry}_{100}\) correspond to the seed entry for no adjacency and full adjacency respectively (see Zamora et al. 2010):

References

Gómez JM (2003). “Spatial patterns in long-distance dispersal of Quercus ilex acorns by jays in a heterogeneous landscape.” Ecography, 26(5), 573--584. doi:10.1034/j.1600-0587.2003.03586.x .

González-Varo JP, López-Bao JV, Guitián J (2013). “Functional diversity among seed dispersal kernels generated by carnivorous mammals.” Journal of Animal Ecology, 82(3), 562--571. doi:10.1111/1365-2656.12024 .

Jordano P, García C, Godoy JA, García-Castaño JL (2007). “Differential contribution of frugivores to complex seed dispersal patterns.” Proceedings of the National Academy of Sciences, 104(9), 3278--3282. doi:10.1073/pnas.0606793104 , http://www.pnas.org/content/104/9/3278.abstract.

Matías L, Zamora R, Mendoza I, Hódar JA (2010). “Seed dispersal patterns by large frugivorous mammals in a degraded mosaic landscape.” Restoration Ecology, 18(5), 619--627. doi:10.1111/j.1526-100X.2008.00475.x .

Pons J, Pausas JG (2007). “Acorn dispersal estimated by radio-tracking.” Oecologia, 153(4), 903--911. doi:10.1007/s00442-007-0788-x .

Zamora R, Hódar JA, Matías L, Mendoza I (2010). “Positive adjacency effects mediated by seed disperser birds in pine plantations.” Ecological Applications, 20(4), 1053--1060. doi:10.1890/09-0055.1 .

Author

Antonio J Pérez-Luque (ajpelu@gmail.com)